changed ssize_t onto int
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@7758 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ motion_planning/sbpl/src/planners/planner.h
@@ -123,7 +123,7 @@ public:
 
   /** \return The number of states expanded during the last replan()
       operation, or -1 if this information is not available. */
-  virtual ssize_t get_n_expands() const { return -1; }
+  virtual int get_n_expands() const { return -1; }
 
 	//setting initial solution eps 
    //This parameter is ignored in planners that don't have a notion of eps
